Рефакторинг: разбиение утилитарного модуля на два модуля: общий и утилиты для карточки фильма
